Refactor OAG God Object into modular package structure
Split the 724-line oag.py into a modular package with core logic in base.py and vendor-specific data in a vendors/ subdirectory. Implemented dynamic aggregation in __init__.py to maintain the public API and backward compatibility.
